requirements.txt
fastapi
uvicorn
python-dotenv
requests
pdfplumber
beautifulsoup4
lxml
sentence-transformers
faiss-cpu
anthropic
